requirements.txt
attrs>=19.3
Babel>=2.8
Click>=7.0
lupa>=1.9
schema>=0.7
requests>=2.23
